# Conflicts:
#	client/src/main/java/client/handlers/NoteSearchResult.java
#	client/src/main/java/client/scenes/NewNoteTitleCtrl.java
#	client/src/main/java/client/scenes/NoteOverviewCtrl.java
#	commons/src/main/java/commons/Note.java

+++ client/src/main/java/client/scenes/NoteOverviewCtrl.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 import java.util.OptionalLong;
 import java.util.ResourceBundle;
 
+import javafx.scene.input.MouseButton;
 import client.handlers.NoteSearchResult;
 import client.services.Markdown;
 import client.utils.NoteUtils;
@@ -20,6 +21,7 @@ import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
+import javafx.scene.input.KeyCode;
 import javafx.scene.web.WebView;
 import javafx.stage.Modality;
 import javafx.scene.input.KeyEvent;
@@ -64,8 +66,15 @@ public class NoteOverviewCtrl implements Initializable {
     private Label selectedNoteTitle;
     @FXML
     private TextArea selectedNoteContent;
+    @FXML
+    private ContextMenu contextMenu;
+    @FXML
+    private MenuItem changeTitle;
+    @FXML
+    private MenuItem refreshNote;
+    @FXML
+    private MenuItem deleteNote;
     private String selectedNoteContentBuffer;
-
     private long selectedNoteId;
 
     @Inject
@@ -119,12 +128,25 @@ public class NoteOverviewCtrl implements Initializable {
         Optional<Note> note = fetchSelectedNote();
         if (note.isEmpty()) return;
         else {
-            server.deleteNote(note.get().getId());
+            String message = "Are you sure you want to delete this note?";
+            String title = "Confirm deletion";
+            int choice = JOptionPane.showConfirmDialog(
+                    null,
+                    message,
+                    title,
+                    JOptionPane.YES_NO_OPTION,
+                    JOptionPane.QUESTION_MESSAGE
+            );
+            if (choice == JOptionPane.YES_OPTION) {
+                server.deleteNote(note.get().getId());
+                selectedNoteTitle.setText(" ");
+                selectedNoteContent.setText(" ");
+                refresh();
+                selectedNoteContent.setDisable(true);
+            } else {
+                refresh();
+            }
         }
-        selectedNoteTitle.setText(" ");
-        selectedNoteContent.setText(" ");
-        refresh();
-        selectedNoteContent.setDisable(true);
     }
 
     public void emptySearchText() {
@@ -173,8 +195,33 @@ public class NoteOverviewCtrl implements Initializable {
         Optional<Note> note = fetchSelectedNote();
         if (note.isEmpty()) return;
 
-        selectedNoteTitle.setText(note.get().title);
-        selectedNoteContent.setText(note.get().content);
+        displaySelectedNote(note.get());
+    }
+
+    /**
+     * Displays the provided note by filling the title and content in the overview
+     *
+     * @param note the note to display
+     */
+    public void displaySelectedNote(Note note) {
+        selectedNoteTitle.setText(note.getTitle());
+        selectedNoteContent.setText(note.getContent());
+    }
+
+    /**
+     * Displays the provided searchResult by selecting its note and text position
+     *
+     * @param searchResult the searchResult to display
+     */
+    public void displaySelectedNote(NoteSearchResult searchResult) {
+        Note note = searchResult.getNote();
+
+        this.selectedNoteId = note.getId();
+        displaySelectedNote(note);
+
+        int startIndex = searchResult.getStartIndex();
+        int endIndex = searchResult.getEndIndex();
+        selectedNoteContent.selectRange(startIndex, endIndex);
     }
 
     /**
@@ -230,10 +277,8 @@ public class NoteOverviewCtrl implements Initializable {
 
         if (text.startsWith("#")) {
             List<NoteSearchResult> foundInNotes = searchNoteContent(text.replaceFirst("#", ""));
-            if (!foundInNotes.isEmpty()) {
-                setViewableNotes(foundInNotes.stream().map(NoteSearchResult::getNoteDTO).toList());
-                mainCtrl.showSearchContent(foundInNotes);
-            }
+            setViewableNotes(foundInNotes.stream().map(NoteSearchResult::getNoteDTO).distinct().toList());
+            mainCtrl.showSearchContent(foundInNotes);
         } else {
             searchAllNotes(text);
         }
@@ -256,7 +301,7 @@ public class NoteOverviewCtrl implements Initializable {
                 if (!foundIndices.isEmpty()) {
                     // Indexes should be recalculated when a user clicks on the note
                     // since the note gets updated and could change!!!
-                    foundIndices.forEach(i -> foundInNotes.add(new NoteSearchResult(n, i)));
+                    foundIndices.forEach(i -> foundInNotes.add(new NoteSearchResult(n, i, queryString.length())));
                 }
             } catch (ProcessOperationException e) {
                 System.out.println(e.getMessage());
@@ -287,11 +332,21 @@ public class NoteOverviewCtrl implements Initializable {
         displaySelectedNote();
     }
 
+    public void showContextMenu() {
+        contextMenu.getItems();
+    }
+
     /**
-     * Currently only has a keyboard shortcut for refreshing/searching
-     * more shortcuts can be added in the future.
+     * Pressed key
+     * ENTER:- refresh
+     * ESCAPE:- sets focus on to the search bar
+     * A:- Shows the window to add a note.
+     * Other shortcuts:
+     * CTRL+T: edits title of a selected note
+     * RIGHT MOUSE CLICK: shows a menu which allows user to delete/refresh/edit a note
+     * currently only works when a user right-clicks on the table and not the individual cells.
      *
-     * @param e Pressed key
+     * @param e
      */
     public void keyPressed(KeyEvent e) {
         switch (e.getCode()) {
@@ -305,7 +360,19 @@ public class NoteOverviewCtrl implements Initializable {
                 addNote();
                 break;
             default:
-                break;
+                if (table.getSelectionModel().getSelectedItem() != null) {
+                    table.setOnMousePressed(event -> {
+                        if (event.getButton() == MouseButton.SECONDARY) {
+                            showContextMenu();
+                        }
+                    });
+                }
+                if (e.getCode() == KeyCode.T && e.isControlDown()) {
+                    setTitle();
+                    break;
+                } else {
+                    break;
+                }
         }
     }
